refactor front-facing API
