+++ dabest/_api.py
@@ -6,7 +6,7 @@
 
 def load(data, idx, x=None, y=None, paired=False, id_col=None, **kwargs):
     '''
-    Create a specialised object for estimation statistics.
+    Loads data in preparation for estimation statistics.
     This is designed to work with pandas DataFrames.
 
     Keywords
@@ -23,10 +23,26 @@ def load(data, idx, x=None, y=None, paired=False, id_col=None, **kwargs):
 
     paired: boolean, default False.
 
-    Examples
+    Example
     --------
-    TO BE ADDED
+    1. Load libraries.
 
+    >>> import numpy as np
+    >>> import pandas as pd
+    >>> import dabest
+
+    2. Create dummy data for demonstration.
+
+    >>> np.random.seed(88888)
+    >>> N = 10
+    >>> c1 = sp.stats.norm.rvs(loc=100, scale=5, size=N)
+    >>> t1 = sp.stats.norm.rvs(loc=115, scale=5, size=N)
+    >>> df = pd.DataFrame({'Control 1' : c1, 'Test 1': t1})
+
+    3. Load the data.
+
+    >>> my_data = dabest.load(df, idx=("Control 1", "Test 1"))
+    
     '''
     from .classes import Dabest
