+++ dabest/_stats_tools/confint_1group.py
