+++ dabest/tests/test_02_plotting.py
@@ -11,7 +11,7 @@ import numpy as np
 import scipy as sp
 import pandas as pd
 import seaborn as sns
-from ..api import load
+from .._api import load
 from .utils import create_dummy_dataset, get_swarm_yspans
